requirements.txt
anthropic
langchain
langchain-community
chromadb
streamlit
sentence-transformers
pypdf
python-dotenv
torch